add safe condition to link creation

+++ recipes/passenger.rb
@@ -40,6 +40,7 @@ node['passenger_nginx_repo']['apps'].each do |app|
   end
   execute "enable vhost" do
     command "ln -s #{File.join(node['nginx']['dir'], 'sites-available', app['appname'])} #{File.join(node['nginx']['dir'], 'sites-enabled', app['appname'])}"
+    not_if { ::File.exists?(File.join(node['nginx']['dir'], 'sites-enabled', app['appname'])) }
     only_if { ::File.exists?(File.join(node['nginx']['dir'], 'sites-available', app['appname'])) }
   end
 end
